feat(content-grid): Fix alignment - add fonts
