+++ .vscode/settings.json
@@ -1,3 +1,3 @@
 {
-  "svg.preview.background": "editor"
+  "svg.preview.background": "white"
 }
