+++ components/RichTextComponent.vue
@@ -1,23 +1,25 @@
 <!-- eslint-disable vue/no-v-html -->
 <template>
-  <div v-if="props.isPreview" class="preview" v-html="previewHtml"></div>
-  <div v-else>
-    <span
-      v-for="(block, index) in htmlContent"
-      :key="index"
-      :class="props.textAlign || 'center'"
-      v-html="block"
-    ></span>
-  </div>
+  <div
+    v-if="props.isPreview && props.content"
+    class="preview"
+    v-html="previewHtml"
+  ></div>
+  <span
+    v-for="(block, index) in htmlContent"
+    v-else
+    :key="index"
+    v-html="block"
+  >
+  </span>
 </template>
 
 <script lang="ts" setup>
 import edjsHTML from 'editorjs-html'
 const props = defineProps<{
-  content: string
+  content: string | null
   isPreview?: boolean
   previewLines?: number
-  textAlign?: 'center' | 'left'
 }>()
 
 const htmlContent = ref<string[] | Error>([])
@@ -26,13 +28,13 @@ const previewHtml = ref<string>('')
 watch(
   () => props.content,
   () => {
+    if (!props.content) return
     htmlContent.value = edjsHTML().parseStrict(JSON.parse(props.content))
     previewHtml.value = htmlContent.value.find((x) => x.includes('<p>')) || ''
   },
   { immediate: true, deep: true }
 )
 </script>
-<!-- TODO: Fix dirty solution .center -->
 <style scoped lang="scss">
 .preview :deep(p) {
   overflow: hidden;
@@ -40,11 +42,8 @@ watch(
   -webkit-line-clamp: v-bind('props.previewLines || 3');
   -webkit-box-orient: vertical;
 }
-.center {
-  display: flex;
-  justify-content: center;
-}
-.left {
-  text-align: left;
+
+.rich-text {
+  display: block;
 }
 </style>
