+++ nuxt.config.ts
@@ -10,6 +10,7 @@ export default defineNuxtConfig({
     '@nuxtjs/color-mode',
     '@nuxt/image',
     'vue3-carousel-nuxt',
+    '@nuxtjs/google-fonts',
   ],
   strapi: {
     url: process.env.STRAPI_URL || 'http://localhost:1337',
@@ -35,4 +36,10 @@ export default defineNuxtConfig({
     preset: 'vercel-edge',
   },
   devtools: { enabled: true },
+  googleFonts: {
+    families: {
+      Lato: [100, 200, 300, 400, 500, 600, 700, 800, 900],
+      Orbitron: [100, 200, 300, 400, 500, 600, 700, 800, 900],
+    },
+  },
 })
