+++ package.json
@@ -46,6 +46,7 @@
     "@nuxtjs/color-mode": "^3.2.0",
     "@nuxtjs/eslint-config-typescript": "^12.0.0",
     "@nuxtjs/eslint-module": "^4.1.0",
+    "@nuxtjs/google-fonts": "^3.1.3",
     "@nuxtjs/stylelint-module": "^5.1.0",
     "@types/node": "^18",
     "@typescript-eslint/eslint-plugin": "^5.59.11",
