+++ pages/abteilung/[slug].vue
@@ -1,32 +1,33 @@
 <template>
-  <div v-if="stepAttributes">
+  <div v-if="stepAttributes" class="content-grid">
     <BaseBanner
       :title="stepAttributes.Name"
       :description="stepAttributes.Description"
       :is-rich-text="true"
       :image="stepAttributes.logo.data"
     />
-    <section
-      v-for="(zone, index) in stepAttributes.pageZone"
-      :key="index"
-      :class="index % 2 === 0 ? 'section' : 'colored-section'"
-    >
-      <SectionComponent
+    <template v-for="(zone, index) in stepAttributes.pageZone" :key="index">
+      <PageZoneSectionComponent
         v-if="zone.__component == 'pages.section'"
-        :title="(zone as Section).Title"
-        :description="(zone as Section).Description"
-        :image="(zone as Section).Image.data[0]"
+        :zone="(zone as Section)"
         :index="index"
-        class="container"
       />
-      <div v-if="zone.__component == 'pages.kastenzeddel'" class="container">
-        <TitleComponent title="Kastenzeddel" :index="index"></TitleComponent>
-        <KastenzeddelComponent
-          :kastenzeddel="(zone as Kastenzeddel)"
-          :step="stepAttributes.Slug"
-        />
-      </div>
-      <div v-if="zone.__component == 'pages.group'" class="container">
+      <StepZoneKastenzeddelComponent
+        v-if="zone.__component == 'pages.kastenzeddel'"
+        :kastenzeddel="(zone as Kastenzeddel)"
+        :step="stepAttributes.Slug"
+        :index="index"
+      />
+      <section
+        v-if="zone.__component == 'pages.group'"
+        :class="[
+          'pt-medium pb-medium',
+          {
+            'full-width content-grid bg-primary-50 dark:bg-primary-800':
+              index % 2 === 1,
+          },
+        ]"
+      >
         <TitleComponent :title="zone.Title" :index="index"></TitleComponent>
         <div v-if="(zone as Group).members" class="member-section">
           <UserCard
@@ -35,44 +36,41 @@
             :user="member"
           ></UserCard>
         </div>
-      </div>
-      <DocumentComponent
+      </section>
+      <PageZoneDocumentComponent
         v-if="zone.__component == 'pages.document'"
         :zone="(zone as Document)"
         :index="index"
-        class="container"
       />
-      <div v-if="zone.__component == 'pages.image'" class="container">
-        <TitleComponent :title="zone.Title" :index="index"></TitleComponent>
-        <ImageSliderComponent
-          v-if="(zone as ImageZone).images"
-          :images="(zone as ImageZone).images.data"
-        />
-      </div>
-    </section>
+      <PageZoneImageComponent
+        v-if="zone.__component == 'pages.image'"
+        :title="zone.Title"
+        :zone="(zone as ImageZone)"
+        :index="index"
+      ></PageZoneImageComponent>
+    </template>
   </div>
 </template>
 <script setup lang="ts">
 const route = useRoute()
 const step = useState<Steps | null>(() => null)
-
 const stepAttributes = computed(() => step.value?.data[0]?.attributes || null)
 
 const fetchData = async () => {
   const { data, error: fetchError } = await getStep(route.params.slug as string)
   if (fetchError.value) {
-    await navigateTo('/not-found')
+    await navigateTo('/404')
     throw createError({
       statusCode: 404,
       statusMessage: fetchError.value.name,
     })
   }
   step.value = data.value
-}
 
-useHead({
-  title: `Pfadi Nünenen - ${stepAttributes.value?.Name || ''}`,
-})
+  useHead({
+    title: `Pfadi Nünenen - ${stepAttributes.value?.Name || ''}`,
+  })
+}
 
 onMounted(fetchData)
 watch(() => route.params.slug, fetchData, { immediate: true })
